requirements.txt
# Google Workspace CLI
google-workspace-cli>=1.0.0

# Slack MCP & utilities
slack-sdk>=3.23.0
slack-bolt>=1.18.0

# Trello API
py-trello>=0.19.0

# Claude API
anthropic>=0.7.0

# Web Search
duckduckgo-search>=3.9.0

# Data models
pydantic>=2.0.0

# Caching & persistence
python-dateutil>=2.8.2
pytz>=2023.3

# Logging & config
python-dotenv>=1.0.0

# Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Optional: async support
httpx>=0.24.0
